+++ zoo/src/test/scala/com/intel/analytics/zoo/pipeline/api/keras/layers/internal/InternalMMSpec.scala
@@ -19,7 +19,7 @@ package com.intel.analytics.zoo.pipeline.api.keras.layers.internal
 import com.intel.analytics.bigdl.nn.{Input, Log}
 import com.intel.analytics.bigdl.tensor.Tensor
 import com.intel.analytics.bigdl.utils.{Shape, T}
-import com.intel.analytics.zoo.pipeline.api.autograd.{AutoGrad, InternalMM, Variable}
+import com.intel.analytics.zoo.pipeline.api.autograd.{AutoGrad, Variable}
 import com.intel.analytics.zoo.pipeline.api.keras.models.Model
 import com.intel.analytics.zoo.pipeline.api.keras.serializer.ModuleSerializationTest
 import org.scalatest.{FlatSpec, Matchers}
